src/3-load/requirements.txt
# ------------------------------------ app\3-load\requirements.txt ------------------------------------ 
pyodbc==4.0.39
pandas==1.5.3
azure-identity==1.12.0
azure-storage-file-datalake==12.13.0
tenacity==8.2.3